fix: Attachment.findBySignature
